requirements.txt
bdflib>=2.1.0
Pillow>=10.1.0
